+++ CHANGELOG.md
@@ -4,6 +4,9 @@ The changelog for [Kommunicate-iOS-SDK](https://github.com/Kommunicate-io/Kommun
 
 ## [Unreleased]
 
+### Enhancements
+- [CM-451] Add support for setting a prefilled message to send before launching a chat.
+
 ## [5.7.0] - 2020-09-10
 
 ### Enhancements